(Fixes issue #11) : USB support With the introduction of IOS58 with USB2.0 support, fatInitDefault doesn't initialize the filesystem anymore. dhegw advised to use libogc r4264.
Updated libogc package to r4264.
Also merged USB keyboard files as the USB subsystem has been changed too. 
(Application was not returning to homebrew channel after shutdown, when USB keyboard was initialized)

Tested that a wbfs formatted fs is not initialized.
Tested on Wii with System menu 4.3E with IOS58
Tested on Wii with System menu 4.0E (unknown IOS)

When starting, both USB and SD card are initialized,
USB is a lot faster than SD card, so USB is preferred (when already installed on both devices)
when both devices are available, but no MSX installation is found, there is a menu to choose from USB or SD.
Then the user is still asked whether to start the installation.

additional modification: added svn:ignore property on V1.0.x\build folder